# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koraspond Microservices API",
        "description": "API documentation for the application endpoints.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "10a7bfb6069f5588a62264143f23fd20",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "testuser8@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userAbilityRules": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "action": {
                                                        "type": "string",
                                                        "example": "manage"
                                                    },
                                                    "subject": {
                                                        "type": "string",
                                                        "example": "all"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "accessToken": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL2FkbWluL2xvZ2luIiwiaWF0IjoxNzI0OTMwNTEzLCJleHAiOjE3MjUwMTY5MTMsIm5iZiI6MTcyNDkzMDUxMywianRpIjoiR0ZkdHc0cFZkMHpJbEFiUCIsInN1YiI6IjExIiwicHJ2IjoiZGY4ODNkYjk3YmQwNWVmOGZmODUwODJkNjg2YzQ1ZTgzMmU1OTNhOSJ9.U-5llflEKkALpISH1g8cOLWaDCpjP_Od2y2TfK6ruQo"
                                        },
                                        "userData": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 11
                                                },
                                                "fullName": {
                                                    "type": "string",
                                                    "example": "Test User"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "/images/avatars/avatar-1.png"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "testuser8@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "permissions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "text": {
                                                        "type": "string",
                                                        "example": "Dashboard Permissions"
                                                    },
                                                    "permissions": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "text": {
                                                                    "type": "string",
                                                                    "example": "Dashboard Filters"
                                                                },
                                                                "ability": {
                                                                    "type": "boolean",
                                                                    "example": true
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid credentials"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/admin/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User registration",
                "operationId": "c1524b62a869fbe19d07b7a83ddc379e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test User"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "testuser8@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userAbilityRules": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "action": {
                                                        "type": "string",
                                                        "example": "manage"
                                                    },
                                                    "subject": {
                                                        "type": "string",
                                                        "example": "all"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "accessToken": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LjAuMC4xOjgwMDAvYXBpL2FkbWluL3JlZ2lzdGVyIiwiaWF0IjoxNzI0ODg0MDk1LCJleHAiOjE3MjQ5NzA0OTUsIm5iZiI6MTcyNDg4NDA5NSwianRpIjoiazE5Um1QNzlNOFlLQzd0OCIsInN1YiI6IjExIiwicHJ2IjoiZGY4ODNkYjk3YmQwNWVmOGZmODUwODJkNjg2YzQ1ZTgzMmU1OTNhOSJ9.kQgMLsl2b4wnMTfUOiDmVxXIrtBZZZOUMACsrrtIymo"
                                        },
                                        "userData": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 11
                                                },
                                                "fullName": {
                                                    "type": "string",
                                                    "example": "Test User"
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "/images/avatars/avatar-1.png"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "testuser8@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "permissions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "text": {
                                                        "type": "string",
                                                        "example": "Dashboard Permissions"
                                                    },
                                                    "permissions": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "text": {
                                                                    "type": "string",
                                                                    "example": "Dashboard Filters"
                                                                },
                                                                "ability": {
                                                                    "type": "boolean",
                                                                    "example": true
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation errors or already registered email"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}